Update MLA implementation
